setup uint64_t instead of unsigned long long

+++ RSA.h
@@ -2,7 +2,9 @@
 
 #define rsa
 
-unsigned long long encrypt(unsigned long long input);
-unsigned long long decrypt(unsigned long long input);
+#include <stdint.h>
+
+uint64_t encrypt(uint64_t input);
+uint64_t decrypt(uint64_t input);
 
 #endif
